k/alter_configs: generalized alter config request handling
`ALTER_CONFIGS` and `INCREMENTAL_ALTER_CONFIGS` Kafka APIs have almost
identical handler logic. Moved `ALTER_CONFIGS` handler logic to utils
file, in order to use already existing implementation in both
cases. All functions were replaced with function templates as request
and responses types are different for both APIs yet they are very
similar.
